+++ src/views/payment-options-view.js
@@ -45,6 +45,12 @@ PaymentOptionsView.prototype._addPaymentOption = function (paymentOptionID) {
     case paymentOptionIDs.paypal:
       html = html.replace(/@ICON/g, 'logoPayPal');
       html = html.replace(/@OPTION_TITLE/g, this.strings.PayPal);
+      html = html.replace(/@OPTION_TITLE/g, this.strings.Card);
+      html = html.replace(/@CLASSNAME/g, '');
+      break;
+    case paymentOptionIDs.paypalCredit:
+      html = html.replace(/@ICON/g, 'logoPayPalCredit');
+      html = html.replace(/@OPTION_TITLE/g, this.strings['PayPal Credit']);
       html = html.replace(/@CLASSNAME/g, '');
       break;
     default: