+++ CHANGELOG.md
@@ -7,6 +7,7 @@ unreleased
 - Use version 3.12.0 of braintree-web
 - Improve loading transition
 - Add full support for IE 9 and 10
+- Add support for PayPal Credit
 
 1.0.0-beta.6
 ------------